Switched API from GET to POST only, added data limits, more security checking
Added MaxByteReader data limit.  Required a switch from GET to POST for
this to work.  Removed regular expression check for URL scheme and used
Golang’s http library to look for a scheme.  Further checks to see if a
URL is a potentially valid URL.  Small bug fix in Javascript that would
tag on a http:// scheme if the given URL scheme wasn’t all lowercase.